+++ tests/test_router_serve_file.c
@@ -12,32 +12,30 @@ enum HSServeFlowResponse _test_serve(struct HSRoute *route, struct HSServeFlowPa
   assert_true(route != NULL);
   assert_true(params != NULL);
 
-  params->response->code              = HS_HTTP_RESPONSE_CODE_NOT_FOUND;
-  params->response->headers->count    = 3;
-  params->response->headers->pairs    = malloc(sizeof(struct HSKeyValue) * params->response->headers->count);
-  params->response->headers->pairs[0] = hs_types_new_key_value(strdup("header1"), strdup("value1"));
-  params->response->headers->pairs[1] = hs_types_new_key_value(strdup("header2"), strdup("value2"));
-  params->response->headers->pairs[2] = hs_types_new_key_value(strdup("header3"), strdup("value3"));
-  params->response->cookies->count    = 3;
-  params->response->cookies->cookies  = malloc(sizeof(struct HSCookie) * params->response->cookies->count);
-  for (size_t index = 0; index < params->response->cookies->count; index++)
-  {
-    params->response->cookies->cookies[index] = hs_types_new_cookie();
-  }
-  params->response->cookies->cookies[0]->name      = strdup("c1");
-  params->response->cookies->cookies[0]->value     = strdup("v1");
-  params->response->cookies->cookies[0]->expires   = strdup("1 1 1980");
-  params->response->cookies->cookies[0]->max_age   = 200;
-  params->response->cookies->cookies[0]->secure    = true;
-  params->response->cookies->cookies[0]->http_only = true;
-  params->response->cookies->cookies[0]->domain    = strdup("mydomain");
-  params->response->cookies->cookies[0]->path      = strdup("/somepath");
-  params->response->cookies->cookies[0]->same_site = HS_COOKIE_SAME_SITE_NONE;
-  params->response->cookies->cookies[1]->name      = strdup("c2");
-  params->response->cookies->cookies[1]->value     = strdup("v2");
-  params->response->cookies->cookies[1]->same_site = HS_COOKIE_SAME_SITE_STRICT;
-  params->response->cookies->cookies[2]->name      = strdup("c3");
-  params->response->cookies->cookies[2]->value     = strdup("v3");
+  params->response->code = HS_HTTP_RESPONSE_CODE_NOT_FOUND;
+  hs_types_key_value_array_add(params->response->headers, strdup("header1"), strdup("value1"));
+  hs_types_key_value_array_add(params->response->headers, strdup("header2"), strdup("value2"));
+  hs_types_key_value_array_add(params->response->headers, strdup("header3"), strdup("value3"));
+  struct HSCookie *cookie = hs_types_new_cookie();
+  cookie->name      = strdup("c1");
+  cookie->value     = strdup("v1");
+  cookie->expires   = strdup("1 1 1980");
+  cookie->max_age   = 200;
+  cookie->secure    = true;
+  cookie->http_only = true;
+  cookie->domain    = strdup("mydomain");
+  cookie->path      = strdup("/somepath");
+  cookie->same_site = HS_COOKIE_SAME_SITE_NONE;
+  hs_types_cookies_add(params->response->cookies, cookie);
+  cookie            = hs_types_new_cookie();
+  cookie->name      = strdup("c2");
+  cookie->value     = strdup("v2");
+  cookie->same_site = HS_COOKIE_SAME_SITE_STRICT;
+  hs_types_cookies_add(params->response->cookies, cookie);
+  cookie        = hs_types_new_cookie();
+  cookie->name  = strdup("c3");
+  cookie->value = strdup("v3");
+  hs_types_cookies_add(params->response->cookies, cookie);
 
   params->response->mime_type = HS_MIME_TYPE_TEXT_PLAIN;