+++ tests/test_router_write_common_response_header.c
@@ -21,8 +21,8 @@ void test_impl()
   hs_types_release_key_value_array(headers);
   hs_types_release_cookies(cookies);
 
-  headers = hs_types_new_key_value_array();
-  cookies = hs_types_new_cookies();
+  headers = hs_types_new_key_value_array(50);
+  cookies = hs_types_new_cookies(1);
   buffer  = hs_router_write_common_response_header(HS_HTTP_RESPONSE_CODE_OK, headers, cookies, true);
   text    = string_buffer_to_string(buffer);
   string_buffer_release(buffer);
@@ -31,35 +31,33 @@ void test_impl()
   hs_types_release_key_value_array(headers);
   hs_types_release_cookies(cookies);
 
-  headers           = hs_types_new_key_value_array();
-  cookies           = hs_types_new_cookies();
-  headers->count    = 3;
-  headers->pairs    = malloc(sizeof(struct HSKeyValue) * headers->count);
-  headers->pairs[0] = hs_types_new_key_value(strdup("header1"), strdup("value1"));
-  headers->pairs[1] = hs_types_new_key_value(strdup("header2"), strdup("value2"));
-  headers->pairs[2] = hs_types_new_key_value(strdup("header3"), strdup("value3"));
-  cookies->count    = 3;
-  cookies->cookies  = malloc(sizeof(struct HSCookie) * cookies->count);
-  for (size_t index = 0; index < cookies->count; index++)
-  {
-    cookies->cookies[index] = hs_types_new_cookie();
-  }
-  cookies->cookies[0]->name      = strdup("c1");
-  cookies->cookies[0]->value     = strdup("v1");
-  cookies->cookies[0]->expires   = strdup("1 1 1980");
-  cookies->cookies[0]->max_age   = 200;
-  cookies->cookies[0]->secure    = true;
-  cookies->cookies[0]->http_only = true;
-  cookies->cookies[0]->domain    = strdup("mydomain");
-  cookies->cookies[0]->path      = strdup("/somepath");
-  cookies->cookies[0]->same_site = HS_COOKIE_SAME_SITE_NONE;
-  cookies->cookies[1]->name      = strdup("c2");
-  cookies->cookies[1]->value     = strdup("v2");
-  cookies->cookies[1]->same_site = HS_COOKIE_SAME_SITE_STRICT;
-  cookies->cookies[2]->name      = strdup("c3");
-  cookies->cookies[2]->value     = strdup("v3");
-  buffer                         = hs_router_write_common_response_header(HS_HTTP_RESPONSE_CODE_FORBIDDEN, headers, cookies, true);
-  text                           = string_buffer_to_string(buffer);
+  headers = hs_types_new_key_value_array(1);
+  cookies = hs_types_new_cookies(1);
+  hs_types_key_value_array_add(headers, strdup("header1"), strdup("value1"));
+  hs_types_key_value_array_add(headers, strdup("header2"), strdup("value2"));
+  hs_types_key_value_array_add(headers, strdup("header3"), strdup("value3"));
+  struct HSCookie *cookie = hs_types_new_cookie();
+  cookie->name      = strdup("c1");
+  cookie->value     = strdup("v1");
+  cookie->expires   = strdup("1 1 1980");
+  cookie->max_age   = 200;
+  cookie->secure    = true;
+  cookie->http_only = true;
+  cookie->domain    = strdup("mydomain");
+  cookie->path      = strdup("/somepath");
+  cookie->same_site = HS_COOKIE_SAME_SITE_NONE;
+  hs_types_cookies_add(cookies, cookie);
+  cookie            = hs_types_new_cookie();
+  cookie->name      = strdup("c2");
+  cookie->value     = strdup("v2");
+  cookie->same_site = HS_COOKIE_SAME_SITE_STRICT;
+  hs_types_cookies_add(cookies, cookie);
+  cookie        = hs_types_new_cookie();
+  cookie->name  = strdup("c3");
+  cookie->value = strdup("v3");
+  hs_types_cookies_add(cookies, cookie);
+  buffer = hs_router_write_common_response_header(HS_HTTP_RESPONSE_CODE_FORBIDDEN, headers, cookies, true);
+  text   = string_buffer_to_string(buffer);
   string_buffer_release(buffer);
   assert_string_equal(text, "HTTP/1.1 403 403\r\n"
                       "header1: value1\r\n"