+++ tests/test_types_release_cookies_full.c
@@ -4,14 +4,13 @@
 
 void test_impl()
 {
-  struct HSCookies *cookies = hs_types_new_cookies();
+  struct HSCookies *cookies = hs_types_new_cookies(1);
 
-  cookies->count   = 3;
-  cookies->cookies = malloc(sizeof(struct HSCookie) * cookies->count);
   for (size_t index = 0; index < cookies->count; index++)
   {
-    cookies->cookies[index]       = hs_types_new_cookie();
-    cookies->cookies[index]->name = stringfn_new_empty_string();
+    struct HSCookie *cookie = hs_types_new_cookie();
+    cookie->name = stringfn_new_empty_string();
+    hs_types_cookies_add(cookies, cookie);
   }
 
   hs_types_release_cookies(cookies);