+++ tests/test_types_release_http_request_full.c
@@ -13,19 +13,17 @@ void test_impl()
   request->authorization = stringfn_new_empty_string();
   request->payload       = hs_types_new_http_request_payload(NULL);
 
-  request->cookies->count   = 3;
-  request->cookies->cookies = malloc(sizeof(struct HSCookie) * request->cookies->count);
-  for (size_t index = 0; index < request->cookies->count; index++)
+  for (size_t index = 0; index < 3; index++)
   {
-    request->cookies->cookies[index]       = hs_types_new_cookie();
-    request->cookies->cookies[index]->name = stringfn_new_empty_string();
+    struct HSCookie *cookie = hs_types_new_cookie();
+    cookie->name  = stringfn_new_empty_string();
+    cookie->value = stringfn_new_empty_string();
+    hs_types_cookies_add(request->cookies, cookie);
   }
 
-  request->headers->count = 5;
-  request->headers->pairs = malloc(sizeof(struct HSKeyValue) * request->headers->count);
-  for (size_t index = 0; index < request->headers->count; index++)
+  for (size_t index = 0; index < 5; index++)
   {
-    request->headers->pairs[index] = hs_types_new_key_value(stringfn_new_empty_string(), stringfn_new_empty_string());
+    hs_types_key_value_array_add(request->headers, stringfn_new_empty_string(), stringfn_new_empty_string());
   }
 
   hs_types_release_http_request(request);