+++ tests/test_types_release_http_response_full.c
@@ -6,19 +6,17 @@ void test_impl()
 {
   struct HSHttpResponse *response = hs_types_new_http_response();
 
-  response->cookies->count   = 3;
-  response->cookies->cookies = malloc(sizeof(struct HSCookie) * response->cookies->count);
-  for (size_t index = 0; index < response->cookies->count; index++)
+  for (size_t index = 0; index < 3; index++)
   {
-    response->cookies->cookies[index]       = hs_types_new_cookie();
-    response->cookies->cookies[index]->name = stringfn_new_empty_string();
+    struct HSCookie *cookie = hs_types_new_cookie();
+    cookie->name  = stringfn_new_empty_string();
+    cookie->value = stringfn_new_empty_string();
+    hs_types_cookies_add(response->cookies, cookie);
   }
 
-  response->headers->count = 5;
-  response->headers->pairs = malloc(sizeof(struct HSKeyValue) * response->headers->count);
-  for (size_t index = 0; index < response->headers->count; index++)
+  for (size_t index = 0; index < 5; index++)
   {
-    response->headers->pairs[index] = hs_types_new_key_value(stringfn_new_empty_string(), stringfn_new_empty_string());
+    hs_types_key_value_array_add(response->headers, stringfn_new_empty_string(), stringfn_new_empty_string());
   }
 
   response->content_string = stringfn_new_empty_string();