+++ tests/test_types_release_key_value_array_empty.c
@@ -3,7 +3,7 @@
 
 void test_impl()
 {
-  struct HSKeyValueArray *key_value_array = hs_types_new_key_value_array();
+  struct HSKeyValueArray *key_value_array = hs_types_new_key_value_array(50);
 
   hs_types_release_key_value_array(key_value_array);
 }