+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Routes support all HTTP methods.
 * 404 route supports all child paths by default.
 * Refactor router flow invocation to enable non response routes to modify final HTTP response.
+* New simplified api for adding headers and cookies
 
 ### v0.2.3 (2021-08-24)